requirements.txt
groq==0.4.2
httpx==0.27.0

fastapi==0.115.0
uvicorn==0.30.6

sentence-transformers==2.3.0
transformers==4.36.2

faiss-cpu==1.7.4
numpy==1.24.4
scikit-learn==1.3.2

pypdf==4.3.1
python-multipart

requests==2.31.0
twilio==9.0.0



feedparser==6.0.11
beautifulsoup4==4.12.3

python-dotenv==1.0.1

torch==2.1.2
sentencepiece==0.1.99
cryptography==41.0.7